+++ src/allmydata/test/common_web.py
@@ -1,3 +1,4 @@
+from future.utils import PY2
 
 import treq
 from twisted.internet.defer import (
@@ -6,19 +7,6 @@ from twisted.internet.defer import (
     returnValue,
 )
 from twisted.web.error import Error
-
-from nevow.context import WebContext
-from nevow.testutil import FakeRequest
-from nevow.appserver import (
-    processingFailed,
-    DefaultExceptionHandler,
-)
-from nevow.inevow import (
-    ICanHandleException,
-    IRequest,
-    IResource as INevowResource,
-    IData,
-)
 
 @inlineCallbacks
 def do_http(method, url, **kwargs):
@@ -31,32 +19,16 @@ def do_http(method, url, **kwargs):
     returnValue(body)
 
 
-def render(resource, query_args):
-    """
-    Render (in the manner of the Nevow appserver) a Nevow ``Page`` or a
-    Twisted ``Resource`` against a request with the given query arguments .
-
-    :param resource: The page or resource to render.
-
-    :param query_args: The query arguments to put into the request being
-        rendered.  A mapping from ``bytes`` to ``list`` of ``bytes``.
-
-    :return Deferred: A Deferred that fires with the rendered response body as
-        ``bytes``.
-    """
-    ctx = WebContext(tag=resource)
-    req = FakeRequest(args=query_args)
-    ctx.remember(DefaultExceptionHandler(), ICanHandleException)
-    ctx.remember(req, IRequest)
-    ctx.remember(None, IData)
-
-    def maybe_concat(res):
-        if isinstance(res, bytes):
-            return req.v + res
-        return req.v
-
-    resource = INevowResource(resource)
-    d = maybeDeferred(resource.renderHTTP, ctx)
-    d.addErrback(processingFailed, req, ctx)
-    d.addCallback(maybe_concat)
-    return d
+if PY2:
+    # We can only use Nevow on Python 2 and Tahoe-LAFS still *does* use Nevow
+    # so prefer the Nevow-based renderer if we can get it.
+    from .common_nevow import (
+        render,
+    )
+else:
+    # However, Tahoe-LAFS *will* use Twisted Web before too much longer so go
+    # ahead and let some tests run against the Twisted Web-based renderer on
+    # Python 3.  Later this will become the only codepath.
+    from .common_tweb import (
+        render,
+    )
